FibHanderのエンドポイントと処理の分離
