+++ internal/handler/fib.go
@@ -4,24 +4,14 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
+
+	"fib_api/internal/service"
 )
 
 func FibHandler(w http.ResponseWriter, r *http.Request) {
 	query := r.URL.Query()
 	nString := query.Get("n")
 	n, _ := strconv.Atoi(nString)
-	ans := Fib(n)
+	ans := service.Fib(n)
 	fmt.Fprint(w, ans)
-}
-
-func Fib(n int) int64 {
-	if n == 1 || n == 2 {
-		return 1
-	} else {
-		var a, b int64 = 1,1 
-		for i:= 3; i <= n; i++ {
-			a, b = b, a+b
-		}
-		return b
-	}
 }
